add `optional` setting to auth_required
`optional` signifies that client will provide authentication if and
only if the server requires authentication.

NATS without TLS only authenticates the client and does not provide
mutal authententication. Accepting that the server does not need
authentication even when we have credentials does therefore not
degrade security.

If mutal authentication is needed, then TLS MUST be used.